resturcutre the application
